+++ game/board.py
@@ -4,6 +4,12 @@ from game.bagtiles import DATA
 from pyrae import dle
 from colorama import Fore, Back, Style, init
 
+DICTIONARY = []
+with open('dictionary', 'r') as dictionary:
+    words = dictionary.readlines()
+for word in words:
+    DICTIONARY.extend(word.split())
+
 class NotInternetConnection(Exception):
     pass
 
@@ -89,10 +95,14 @@ class Board():
         return word_multiplier, points
     
     def rae_search(self, word):
-        rae = dle.search_by_word(word)
-        if rae is None:
-            raise NotInternetConnection
-        return 'D' != rae.title[:1]
+        word = word.lower()
+        if word in DICTIONARY:
+            return True
+        else:
+            rae = dle.search_by_word(word)
+            if rae is None:
+                raise NotInternetConnection
+            return 'D' != rae.title[:1]
 
     def is_empty(self):
         return self.grid[7][7].tile is None
@@ -107,35 +117,36 @@ class Board():
             row += 1 if not horizontal else 0
         return False
 
-    def validate_side_cell(self, parameters, validators, cell, index_increment):
-        word, pos, i = parameters[0], parameters[1], parameters[2]
-        is_valid, intersections = validators[0], validators[1]
-        grid, word2, index = self.grid, word[i], 1
+    def validate_side_cell(self, parameters, cell, index_increment):
+        letter, pos, horizontal = parameters[0], parameters[1], parameters[2]
+        grid, side_word, index = self.grid, letter, 1
         while cell:
-            word2 += cell.letter.lower()
-            cell = grid[pos[0] + i + index_increment * index][pos[1] + i].tile
+            side_word += cell.letter.lower()
             index += 1
-        word2_is_valid = self.rae_search(word2)
-        if not word2_is_valid:
+            cell = grid[pos[0]+(index*index_increment)][pos[1]].tile if horizontal else grid[pos[0]][pos[1]+(index*index_increment)].tile
+        return side_word
+
+    def check_cells(self, cells, parameters, validators):
+        cell, sidecell, invertsidecell = cells[0], cells[1], cells[2]
+        letter, pos, horizontal = parameters[0], parameters[1], parameters[2]
+        is_valid, intersections = validators[0], validators[1]
+        if cell:
+            intersections += 1
+            is_valid += 1 if cell.letter == letter.upper() else 0
+            return [is_valid, intersections]
+        elif sidecell and invertsidecell:
+            side_word = self.validate_side_cell([letter, pos, horizontal], invertsidecell, -1)[::-1]
+            side_word += self.validate_side_cell([letter, pos, horizontal], sidecell, 1)[1:]
+        elif sidecell:
+            side_word = self.validate_side_cell([letter, pos, horizontal], sidecell, 1)
+        elif invertsidecell:
+            side_word = self.validate_side_cell([letter, pos, horizontal], invertsidecell, -1)[::-1]
+        side_word_is_valid = self.rae_search(side_word)
+        if not side_word_is_valid:
             is_valid = -9999
         else:
             is_valid += 1
             intersections += 1
-        return [is_valid, intersections]
-
-    def check_cells(self, cells, parameters, validators):
-        cell, sidecell, invertsidecell = cells[0], cells[1], cells[2]
-        word, pos, i = parameters[0], parameters[1], parameters[2]
-        is_valid, intersections = validators[0], validators[1]
-        if cell:
-            intersections += 1
-            is_valid += 1 if cell.letter == word[i].upper() else 0
-        elif sidecell:
-            result = self.validate_side_cell([word, pos, i], [is_valid, intersections], sidecell, 1)
-            is_valid, intersections = result[0], result[1]
-        elif invertsidecell:
-            result = self.validate_side_cell([word, pos, i], [is_valid, intersections], invertsidecell, -1)
-            is_valid, intersections = result[0], result[1]
         return [is_valid, intersections]
 
     def validate_not_empty(self, word, pos, horizontal):
@@ -143,11 +154,18 @@ class Board():
         is_valid = 0
         grid = self.grid
         for i in range(len(word)):
-            cell = grid[pos[0] + (i if not horizontal else 0)][pos[1] + (i if horizontal else 0)].tile
-            sidecell = grid[pos[0] + i][pos[1] + 1].tile if not horizontal else grid[pos[0] + 1][pos[1] + i].tile
-            invertsidecell = grid[pos[0] + i][pos[1] - 1].tile if not horizontal else grid[pos[0] - 1][pos[1] + i].tile
-            checked = self.check_cells([cell, sidecell, invertsidecell], [word, pos, i], [is_valid, intersections])
-            is_valid, intersections = checked[0], checked[1]
+            if horizontal:
+                cell = grid[pos[0]][pos[1]+i].tile
+                sidecell = grid[pos[0] + 1][pos[1] + i].tile
+                invertsidecell = grid[pos[0] - 1][pos[1] + i].tile
+            else:
+                cell = grid[pos[0]+i][pos[1]].tile
+                sidecell = grid[pos[0] + i][pos[1] + 1].tile
+                invertsidecell = grid[pos[0] + i][pos[1] - 1].tile
+            if cell or sidecell or invertsidecell:
+                location = (pos[0],pos[1]+i) if horizontal else (pos[0]+i,pos[1])
+                checked = self.check_cells([cell, sidecell, invertsidecell], [word[i], location, horizontal], [is_valid, intersections])
+                is_valid, intersections = checked[0], checked[1]
         return is_valid != 0 and is_valid == intersections
 
     def validate(self, word, pos, horizontal):